Rename asset_helpers.py to normalizer.py and move it under sdk/definitions/asset.
